requirements.txt
asgiref==3.2.5
asn1crypto==1.3.0
bcrypt==3.1.7
cffi
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-bootstrap3==12.0.3
django-mail-templated==2.6.5
django-widget-tweaks==1.4.8
fabric==2.5.0
gunicorn==20.0.4
idna==2.9
invoke==1.4.1
#mysqlclient==1.4.6
paramiko==2.7.1
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
